requirement.txt
numpy==2.4.4
pandas==3.0.2
psycopg2== 2.9.11
python-dotenv== 1.2.2
scikit-learn == 1.8.0